Update 20250709_bkjn_15681(트리와 쿼리).cpp

+++ BKJN/트리/20250709_bkjn_15681(트리와 쿼리).cpp
@@ -13,7 +13,7 @@ void DFS(int vertex, int parent) {
 
 	visited[vertex] = true;
 
-	for (long unsigned int i = 0; i < v[vertex].size(); i++) {
+	for (int i = 0; i < v[vertex].size(); i++) {
 		DFS(v[vertex][i], vertex);
 	}
